# 8313005: Ensure native access check can fold away