This patch helps to ensure that native access checks done by restricted methods can completely fold away (see JBS for details).

This is accomplished by:
1. Marking the `module` field of `java.lang.Class` `@Stable`
2. Using a constant `JavaLangAccess` instance.

I've also added 2 new benchmarks that show the improvement (in two separate classes so that the benchmark methods in each class are comparable to each other in terms of numbers).

Numbers on my machine are as follows:

Before:
```
Benchmark                       Mode  Cnt  Score   Error  Units
MemorySegmentGetUnsafe.panama   avgt   30  0.817 ± 0.002  ns/op
MemorySegmentGetUnsafe.unsafe   avgt   30  0.400 ± 0.003  ns/op

MemorySegmentCopyUnsafe.panama  avgt   30  8.145 ± 0.016  ns/op
MemorySegmentCopyUnsafe.unsafe  avgt   30  7.744 ± 0.012  ns/op
```

After:
```
Benchmark                       Mode  Cnt  Score   Error  Units
MemorySegmentGetUnsafe.panama   avgt   30  0.411 ± 0.002  ns/op
MemorySegmentGetUnsafe.unsafe   avgt   30  0.404 ± 0.003  ns/op

MemorySegmentCopyUnsafe.panama  avgt   30  7.737 ± 0.013  ns/op
MemorySegmentCopyUnsafe.unsafe  avgt   30  7.728 ± 0.011  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Change must be properly reviewed (1 review required, with at least 1 [Committer](https://openjdk.org/bylaws#committer))

### Issue
 * [JDK-8313005](https://bugs.openjdk.org/browse/JDK-8313005): Ensure native access check can fold away (**Enhancement** - P3)


### Reviewers
 * [Maurizio Cimadamore](https://openjdk.org/census#mcimadamore) (@mcimadamore - Committer)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/panama-foreign.git pull/851/head:pull/851` \
`$ git checkout pull/851`

Update a local copy of the PR: \
`$ git checkout pull/851` \
`$ git pull https://git.openjdk.org/panama-foreign.git pull/851/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 851`

View PR using the GUI difftool: \
`$ git pr show -t 851`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/panama-foreign/pull/851.diff">https://git.openjdk.org/panama-foreign/pull/851.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/panama-foreign/pull/851#issuecomment-1650368288)